refactor: Extract plugin; export plugin as default
BREAKING
